# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/intervista": {
      "post": {
        "operationId": "IntervistaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntervistaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interview scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseIntervistaDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or schedule conflict detected"
          },
          "404": {
            "description": "Candidate, interviewer, or ricerca not found"
          }
        },
        "summary": "Schedule a new interview with collision detection",
        "tags": [
          "Interviste"
        ]
      },
      "get": {
        "operationId": "IntervistaController_findAll",
        "parameters": [
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "inizio",
                  "fine",
                  "stato",
                  "candidato",
                  "intervistatori",
                  "ricerca"
                ]
              }
            }
          },
          {
            "name": "fullRelations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "stato",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "candidato",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "intervistatori",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "ricerca",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching interviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseIntervistaDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Retrieve interviews with filters, date ranges, and pagination",
        "tags": [
          "Interviste"
        ]
      }
    },
    "/intervista/{id}": {
      "get": {
        "operationId": "IntervistaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "inizio",
                  "fine",
                  "stato",
                  "candidato",
                  "intervistatori",
                  "ricerca"
                ]
              }
            }
          },
          {
            "name": "fullRelations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interview found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseIntervistaDto"
                }
              }
            }
          },
          "404": {
            "description": "Interview not found"
          }
        },
        "summary": "Get a single interview by ID",
        "tags": [
          "Interviste"
        ]
      },
      "patch": {
        "operationId": "IntervistaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntervistaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interview updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseIntervistaDto"
                }
              }
            }
          },
          "400": {
            "description": "Schedule conflict or validation error"
          },
          "404": {
            "description": "Interview or referenced entity not found"
          }
        },
        "summary": "Update an interview (re-validates schedule conflicts)",
        "tags": [
          "Interviste"
        ]
      },
      "delete": {
        "operationId": "IntervistaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interview deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseIntervistaDto"
                }
              }
            }
          },
          "404": {
            "description": "Interview not found"
          }
        },
        "summary": "Cancel/delete an interview by ID",
        "tags": [
          "Interviste"
        ]
      }
    },
    "/intervista/{id}/intervistatori": {
      "post": {
        "operationId": "IntervistaController_addIntervistatore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtenteIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interviewer added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseIntervistaDto"
                }
              }
            }
          },
          "400": {
            "description": "Interviewer is already busy during interview time"
          },
          "404": {
            "description": "Interview or user not found"
          }
        },
        "summary": "Add an interviewer to an interview",
        "tags": [
          "Interviste"
        ]
      }
    },
    "/intervista/{id}/intervistatori/{utenteId}": {
      "delete": {
        "operationId": "IntervistaController_removeIntervistatore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "utenteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interviewer removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseIntervistaDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot remove if zero interviewers would remain"
          },
          "404": {
            "description": "Interview not found or user is not an interviewer"
          }
        },
        "summary": "Remove an interviewer from an interview",
        "tags": [
          "Interviste"
        ]
      }
    },
...
  },
  "info": {
    "title": "IntervistaServer API",
    "description": "API specification for IntervistaServer: candidate management, recruitment searches, and conflict-validated interview scheduling.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateIntervistaDto": {
        "type": "object",
        "properties": {
          "candidato": {
            "type": "number",
            "minimum": 0
          },
          "intervistatori": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 0
            }
          },
          "ricerca": {
            "type": "number",
            "minimum": 0
          },
          "stato": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "inizio": {
            "format": "date-time",
            "type": "string"
          },
          "fine": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "candidato",
          "intervistatori",
          "ricerca",
          "inizio",
          "fine"
        ]
      },
      "ResponseIntervistaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "candidato": {
            "type": "object"
          },
          "intervistatori": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "ricerca": {
            "type": "object"
          },
          "stato": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "inizio": {
            "format": "date-time",
            "type": "string"
          },
          "fine": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidato",
          "intervistatori",
          "ricerca",
          "inizio",
          "fine"
        ]
      },
      "UpdateIntervistaDto": {
        "type": "object",
        "properties": {
          "stato": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "inizio": {
            "format": "date-time",
            "type": "string"
          },
          "fine": {
            "format": "date-time",
            "type": "string"
          },
          "candidato": {
            "type": "number",
            "minimum": 0
          },
          "intervistatori": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 0
            }
          },
          "ricerca": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "UtenteIdDto": {
        "type": "object",
        "properties": {
          "utenteId": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "utenteId"
        ]
      },
...
    }
  }
}